Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() 
 {
-    float a,b,c; 
-    scanf("%f%f%f",&a,&b,&c);
+    int a,b,c; 
+    scanf("%d%d%d",&a,&b,&c);
     if((a+b>c)&&(b+c>a)&&(a+c>b)){
         printf("Valid");
     }else{
